+++ crates/client/db/src/l1_handler_tx_fee.rs
@@ -4,7 +4,7 @@ use parity_scale_codec::Encode;
 use starknet_api::hash::StarkFelt;
 use starknet_api::transaction::Fee;
 
-use crate::{Column, DbError, DB, DatabaseExt};
+use crate::{Column, DatabaseExt, DbError, DB};
 
 pub struct L1HandlerTxFeeDb {
     pub(crate) db: Arc<DB>,
@@ -19,7 +19,7 @@ impl L1HandlerTxFeeDb {
     pub fn store_fee_paid_for_l1_handler_tx(&self, tx_hash: StarkFelt, fee: Fee) -> Result<(), DbError> {
         let column = self.db.get_column(Column::L1HandlerPaidFee);
 
-        self.db.put_cf(&column, &tx_hash.encode(), &fee.0.to_le_bytes())?;
+        self.db.put_cf(&column, tx_hash.encode(), fee.0.to_le_bytes())?;
         Ok(())
     }
 
@@ -27,7 +27,7 @@ impl L1HandlerTxFeeDb {
     pub fn get_fee_paid_for_l1_handler_tx(&self, tx_hash: StarkFelt) -> Result<Fee, DbError> {
         let column = self.db.get_column(Column::L1HandlerPaidFee);
 
-        if let Some(bytes) = self.db.get_cf(&column, &tx_hash.encode())? {
+        if let Some(bytes) = self.db.get_cf(&column, tx_hash.encode())? {
             let mut buff = [0u8; 16];
 
             buff.copy_from_slice(&bytes);