+++ crates/client/db/src/sierra_classes_db.rs
@@ -4,7 +4,7 @@ use parity_scale_codec::{Decode, Encode};
 use starknet_api::api_core::ClassHash;
 use starknet_api::state::ContractClass;
 
-use crate::{Column, DbError, DB, DatabaseExt};
+use crate::{Column, DatabaseExt, DbError, DB};
 
 /// Allow interaction with the sierra classes db
 pub struct SierraClassesDb {
@@ -19,7 +19,7 @@ impl SierraClassesDb {
     pub fn store_sierra_class(&self, class_hash: ClassHash, class: ContractClass) -> Result<(), DbError> {
         let column = self.db.get_column(Column::SierraContractClasses);
 
-        self.db.put_cf(&column, &class_hash.encode(), &class.encode())?;
+        self.db.put_cf(&column, class_hash.encode(), class.encode())?;
         Ok(())
     }
 
@@ -28,7 +28,7 @@ impl SierraClassesDb {
 
         let opt_contract_class = self
             .db
-            .get_cf(&column, &class_hash.encode())?
+            .get_cf(&column, class_hash.encode())?
             .map(|raw| ContractClass::decode(&mut &raw[..]))
             .transpose()?;