+++ crates/client/sync/src/l2.rs
@@ -155,8 +155,7 @@ where
     if first_block == 1 {
         let state_update =
             provider.get_state_update(BlockId::Number(0)).await.expect("getting state update for genesis block");
-        verify_l2(0, &state_update, overrides, None)
-            .expect("verifying genesis block");
+        verify_l2(0, &state_update, overrides, None).expect("verifying genesis block");
     }
 
     let fetch_stream = (first_block..).map(|block_n| {